Menambahkan schedule, membuat partials action datatables, membuat fitur mass destroy

+++ app/Http/Controllers/Admin/LevelController.php
@@ -4,8 +4,12 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\LevelRequest;
+use App\Http\Requests\Admin\MassDestroyLevelRequest;
 use App\Models\Level;
+use App\Models\Permission;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 use Yajra\DataTables\Facades\DataTables;
 
 class LevelController extends Controller
@@ -17,19 +21,24 @@ class LevelController extends Controller
      */
     public function index(Request $request)
     {
+        abort_if(Gate::denies('level_access'), Response::HTTP_FORBIDDEN, 'Forbidden');
+
         if($request->ajax()){   
             $levels = Level::get();
             return DataTables::of($levels)
-                    ->addColumn('action', function($levels){
-                        $button = '<a href='. route("admin.levels.edit", $levels->id).' class="btn btn-success btn-sm mr-2">edit</a>';
-                        $button .= '
-                            <form action='.route("admin.levels.destroy", $levels->id).' method="POST" class="d-inline-block form-delete">
-                                '. csrf_field() .'
-                                '. method_field("DELETE") .'
-                                <button type="submit" class="btn btn-danger btn-sm btn-delete">delete</button>
-                            </form>
-                        ';
-                        return $button;
+                    ->addColumn('action', function($row){
+                        $showGate       = '';
+                        $editGate       = 'level_edit';
+                        $deleteGate     = 'level_delete';
+                        $crudRoutePart  = 'levels';
+                        
+                        return view('partials.datatables-action', compact(
+                            'showGate', 
+                            'editGate', 
+                            'deleteGate',
+                            'crudRoutePart',
+                            'row',
+                        ));
                     })
                     ->toJson();
         }
@@ -44,7 +53,9 @@ class LevelController extends Controller
      */
     public function create()
     {
-        return view('pages.admin.level.create');
+        abort_if(Gate::denies('level_create'), Response::HTTP_FORBIDDEN, 'Forbidden');
+        $permissions = Permission::all();
+        return view('pages.admin.level.create', compact('permissions'));
     }
 
     /**
@@ -55,7 +66,9 @@ class LevelController extends Controller
      */
     public function store(LevelRequest $request)
     {
-        Level::create($request->all());
+        $level = Level::create($request->all());
+        $level->permissions()->sync($request->input('permissions', []));
+
         return redirect()->route('admin.level.index')->withSuccess('Level berhasil ditambahkan!');
     }
 
@@ -78,7 +91,9 @@ class LevelController extends Controller
      */
     public function edit(Level $level)
     {
-        return view('pages.admin.level.edit', compact('level'));
+        abort_if(Gate::denies('level_edit'), Response::HTTP_FORBIDDEN, 'Forbidden');
+        $permissions = Permission::all();
+        return view('pages.admin.level.edit', compact('level', 'permissions'));
     }
 
     /**
@@ -90,7 +105,9 @@ class LevelController extends Controller
      */
     public function update(LevelRequest $request, Level $level)
     {
-        $level->update();
+        $level->update($request->all());
+        $level->permissions()->sync($request->input('permissions', []));
+
         return redirect()->route('admin.level.index')->withSuccess('Level berhasil diubah!');
     }
 
@@ -102,7 +119,23 @@ class LevelController extends Controller
      */
     public function destroy(Level $level)
     {
+        abort_if(Gate::denies('level_delete'), Response::HTTP_FORBIDDEN, 'Forbidden');
         $level->delete();
         return redirect()->route('admin.level.index')->withSuccess('Level berhasil dihapus!');
+    }
+
+    public function massDestroy(MassDestroyLevelRequest $request)
+    {
+        abort_if(Gate::denies('level_delete'), Response::HTTP_FORBIDDEN, 'Forbidden');
+        $levels = Level::whereIn('id', request('ids'))->get();
+        foreach($levels as $level){
+            if($level-> id === 1){
+                alert()->error('Admin tidak dapat dihapus!');
+                return;
+            }
+            $level->delete();
+        }
+
+        return redirect()->route('admin.levels.index')->withSuccess('Data level(s) berhasil dihapus!');
     }
 }
